feat(runtime): reorganize loaders and deprecate old imports for clarity (#35)
Establishes xulcan/runtime/loaders/ as the dedicated bounded context for YAML ingestion, schema compatibility, and declarative artifact loading. This formalizes the architectural distinction between ontology layers (pure schemas) and runtime ingestion boundaries (IO/materialization).
