+++ app/xulcan/runtime/__init__.py
@@ -1,5 +1,5 @@
 from xulcan.runtime.topology import ResolvedInfrastructure
-from xulcan.runtime.resolver import ManifestResolver
+from xulcan.runtime.loaders import ManifestResolver
 from xulcan.runtime.assembler import RuntimeAssembler
 from xulcan.runtime.context import RuntimeContext
 from xulcan.runtime.llm_registry_adapter import RuntimeLLMRegistryAdapter
