+++ app/xulcan/runtime/loaders/__init__.py
@@ -1,0 +1,15 @@
+"""Runtime loaders — Materialization and ingestion boundaries.
+
+This module contains the runtime ingestion pipeline:
+- ManifestResolver: materializes declarative infrastructure manifests
+- BlueprintLoader: hydrates agent blueprints from YAML
+- Compatibility: schema migrations and adapter transforms
+"""
+
+from xulcan.runtime.loaders.manifest_resolver import ManifestResolver
+from xulcan.runtime.loaders.blueprint_loader import BlueprintLoader
+
+__all__ = [
+    "ManifestResolver",
+    "BlueprintLoader",
+]
