+++ app/xulcan/runtime/loaders/compatibility.py
@@ -1,0 +1,5 @@
+"""Future boundary for schema migrations and compatibility adapters.
+
+This module will handle automatic migrations (e.g., Blueprint v1 -> v2),
+provider alias rewrites, and deprecation transforms.
+"""
